04-calc.md: Remove this topic
This falls under the "fancy query" category that I thing was slowing
the lesson down.  I've collected all the SQL introduced by this lesson
and added SQLite doc pointers for those ideas to the listing in
03-filter.md.  That will give learners a taste of what's possible, and
the SQLite docs will fill them in on any details they'd need.
Dropping this topic allows us to spend more time focusing on the big
picture of SQL, since that's the most important idea I think we can
communicate in a short lesson.