Shifting the join and schema-design topics forward
The first three topics are all about single-table SELECTs.  That's
important, but it's not the most mind-bending part of SQL.  Give folks
an exciting break from single-table SELECTs (before we get back to
them to discuss NULLs and aggregation) by taking some time to discuss
cross-table interactions and how you pick a table layout.  This gives
us a change to argue for DRY practices [1] while folks still have
enough energy to absorb them.  Then we can relax and coast out on the
finer points of SELECT.

[1]: http://en.wikipedia.org/wiki/Don%27t_repeat_yourself