04-join.md: Drop clarification between WHERE and ON
The distinction only matters for outer joins [1], and we never discuss
outer joins.  Here's an example of the distinction for outer joins:

  sqlite> SELECT Site.name, Visited.site, Visited.ident, Visited.dated
     ...> FROM Site
     ...> LEFT OUTER JOIN Visited ON Site.name=Visited.site AND Visited.dated IS NULL;
  DR-1|||
  DR-3|DR-3|752|
  MSK-4|||
  sqlite> SELECT Site.name, Visited.site, Visited.ident, Visited.dated
     ...> FROM Site
     ...> LEFT JOIN Visited ON Site.name=Visited.site
     ...> WHERE Visited.dated IS NULL;
  DR-3|DR-3|752|

So the difference (where it matters for outer joins) is that a left
outer join (for example) will return all rows from the left table even
if the ON clause never matches [2], while all returned rows must match
the WHERE clause.  This is a semantic difference in the *meaning* of
the filtering and has nothing to do with when the filtering happens.

[1]: http://stackoverflow.com/a/354094
[2]: http://en.wikipedia.org/wiki/Join_%28SQL%29#Left_outer_join