06-create.md: Change 'queries' -> 'commands' for create, insert, etc.
That's the term used by the SQLite docs for CREATE TABLE [1].  And I
think it makes sense to use a different term for commands that alter
the database, since 'query' sounds like data *extraction* to me.  I
don't know if my feelings here are backed up by the SQL spec, but this
seems like a simple enough change that that probably doesn't matter.

[1]: https://www.sqlite.org/lang_createtable.html